curl op error handling

+++ src/HCFS/hcfscurl.c
@@ -12,7 +12,7 @@
 
 /* Implemented routines to parse http return code to distinguish errors
 	from normal ops*/
-/*TODO: Retry mechanism if HTTP operations failed due to timeout */
+
 #include "hcfscurl.h"
 
 #include <stdio.h>
@@ -27,6 +27,9 @@
 #include "b64encode.h"
 #include "params.h"
 #include "logger.h"
+#include "macro.h"
+
+#define MAX_RETRIES 5
 
 extern SYSTEM_CONF_STRUCT system_config;
 
@@ -41,21 +44,25 @@ typedef struct {
 * Function name: write_file_function
 *        Inputs: void *ptr, size_t size, size_t nmemb, void *fstream
 *       Summary: Same as fwrite but will return total bytes.
-*  Return value: Total size for writing to fstream
+*  Return value: Total size for writing to fstream. Return 0 if error.
 *
 *************************************************************************/
 size_t write_file_function(void *ptr, size_t size, size_t nmemb, void *fstream)
 {
-	size_t total_size;
+	size_t ret_size;
+	int errcode;
 
-	total_size = fwrite(ptr, size, nmemb, fstream);
+	FWRITE(ptr, size, nmemb, fstream);
 
-	return total_size*size;
+	return ret_size*size;
+
+errcode_handle:
+	return 0;
 }
 
 /************************************************************************
 *
-* Function name: parse_auth_header
+* Function name: parse_swift_auth_header
 *        Inputs: FILE *fptr
 *       Summary: Parse the HTTP header for auth requests and return HTTP
 *                return code (in integer). If successful, auth string is
@@ -63,69 +70,94 @@ size_t write_file_function(void *ptr, size_t size, size_t nmemb, void *fstream)
 *  Return value: Return code from HTTP header, or -1 if error.
 *
 *************************************************************************/
-int parse_auth_header(FILE *fptr)
+int parse_swift_auth_header(FILE *fptr)
 {
 	char httpcode[20], retcode[20], retstatus[20];
+	char *endptr;
 	char temp_string[1024], temp_string2[1024];
-	int ret_val, retcodenum;
+	long ret_num;
+	int retcodenum, ret_val;
+	int ret, errcode;
 
-	fseek(fptr, 0, SEEK_SET);
-	ret_val = fscanf(fptr, "%s %s %s\n", httpcode, retcode, retstatus);
+	FSEEK(fptr, 0, SEEK_SET);
+	ret_val = fscanf(fptr, "%19s %19s %19s\n",
+			httpcode, retcode, retstatus);
 	if (ret_val < 3)
 		return -1;
 
-	retcodenum = atoi(retcode);
+	ATOL(retcode);
+	retcodenum = (int) ret_num;
 
 	if ((retcodenum < 200) || (retcodenum > 299))
 		return retcodenum;
 
-	ret_val = fscanf(fptr, "%s %s\n", temp_string, swift_url_string);
+	ret_val = fscanf(fptr, "%1023s %1023s\n", temp_string,
+			swift_url_string);
 
 	if (ret_val < 2)
 		return -1;
 
-	ret_val = fscanf(fptr, "%s %s\n", temp_string, temp_string2);
+	ret_val = fscanf(fptr, "%1023s %1023s\n", temp_string,
+			temp_string2);
 	if (ret_val < 2)
 		return -1;
 
-	sprintf(swift_auth_string, "%s %s", temp_string, temp_string2);
+	sprintf(swift_auth_string, "%1023s %1023s", temp_string,
+			temp_string2);
 
 	return retcodenum;
+
+errcode_handle:
+	return -1;
 }
 
 /************************************************************************
 *
-* Function name: parse_list_header
+* Function name: parse_swift_list_header
 *        Inputs: FILE *fptr
 *       Summary: Parse the HTTP header for Swift list requests and return HTTP
 *                return code (in integer).
 *  Return value: Return code from HTTP header, or -1 if error.
 *
 *************************************************************************/
-int parse_list_header(FILE *fptr)
+int parse_swift_list_header(FILE *fptr)
 {
 	char httpcode[20], retcode[20], retstatus[20];
 	char temp_string[1024], temp_string2[1024];
 	int ret_val, retcodenum, total_objs;
+	int ret, errcode;
+	long ret_num;
+	char *endptr, *tmpptr;
 
-	fseek(fptr, 0, SEEK_SET);
-	ret_val = fscanf(fptr, "%s %s", httpcode, retcode);
+	FSEEK(fptr, 0, SEEK_SET);
+	ret_val = fscanf(fptr, "%19s %19s", httpcode, retcode);
 	if (ret_val < 2)
 		return -1;
 
-	fgets(retstatus, 19, fptr);
-	retcodenum = atoi(retcode);
+	tmpptr = fgets(retstatus, 19, fptr);
+	if (tmpptr == NULL) {
+		write_log(0, "Error parsing in %s\n", __func__);
+		return -1;
+	}
+	ATOL(retcode);
+	retcodenum = (int) ret_num;
 
 	if ((retcodenum < 200) || (retcodenum > 299))
 		return retcodenum;
 
 	while (!feof(fptr)) {
-		fgets(temp_string, 1000, fptr);
+		tmpptr = fgets(temp_string, 1000, fptr);
+		if (tmpptr == NULL) {
+			write_log(0, "Error parsing in %s\n", __func__);
+			return -1;
+		}
+
 		if (!strncmp(temp_string, "X-Container-Object-Count",
 			sizeof("X-Container-Object-Count")-1)) {
 			ret_val = sscanf(temp_string,
 				"X-Container-Object-Count: %s\n", temp_string2);
-			total_objs = atoi(temp_string2);
+			ATOL(temp_string2);
+			total_objs = (int) ret_num;
 
 			write_log(10, "total objects %d\n", total_objs);
 
@@ -133,6 +165,8 @@ int parse_list_header(FILE *fptr)
 		}
 		memset(temp_string, 0, 1000);
 	}
+	return -1;
+errcode_handle:
 	return -1;
 }
 
@@ -149,16 +183,26 @@ int parse_S3_list_header(FILE *fptr)
 {
 	char httpcode[20], retcode[20], retstatus[20];
 	int ret_val, retcodenum;
+	long ret_num;
+	int errcode;
+	char *endptr, *tmpptr;
 
 	fseek(fptr, 0, SEEK_SET);
-	ret_val = fscanf(fptr, "%s %s", httpcode, retcode);
+	ret_val = fscanf(fptr, "%19s %19s", httpcode, retcode);
 	if (ret_val < 2)
 		return -1;
 
-	fgets(retstatus, 19, fptr);
-	retcodenum = atoi(retcode);
+	tmpptr = fgets(retstatus, 19, fptr);
+	if (tmpptr == NULL) {
+		write_log(0, "Error parsing in %s\n", __func__);
+		return -1;
+	}
+	ATOL(retcode);
+	retcodenum = (int) ret_num;
 
 	return retcodenum;
+errcode_handle:
+	return -1;
 }
 
 /************************************************************************
@@ -174,15 +218,23 @@ int parse_http_header_retcode(FILE *fptr)
 {
 	char httpcode[20], retcode[20], retstatus[20];
 	int ret_val, retcodenum;
+	long ret_num;
+	int errcode;
+	char *endptr;
 
 	fseek(fptr, 0, SEEK_SET);
-	ret_val = fscanf(fptr, "%s %s %s\n", httpcode, retcode, retstatus);
+	ret_val = fscanf(fptr, "%19s %19s %19s\n", httpcode, retcode,
+			retstatus);
 	if (ret_val < 3)
 		return -1;
 
-	retcodenum = atoi(retcode);
+	ATOL(retcode);
+	retcodenum = (int) ret_num;
 
 	return retcodenum;
+
+errcode_handle:
+	return -1;
 }
 
 /************************************************************************
@@ -200,7 +252,7 @@ void dump_list_body(FILE *fptr)
 
 	fseek(fptr, 0, SEEK_SET);
 	while (!feof(fptr)) {
-		ret_val = fscanf(fptr, "%s\n", temp_string);
+		ret_val = fscanf(fptr, "%1023s\n", temp_string);
 		if (ret_val < 1)
 			break;
 		write_log(10, "%s\n", temp_string);
@@ -251,11 +303,19 @@ int hcfs_get_auth_swift(char *swift_user, char *swift_pass, char *swift_url,
 	FILE *fptr;
 	CURL *curl;
 	char filename[100];
+	int errcode, ret;
+	int num_retries;
 
 	sprintf(filename, "/run/shm/swiftauth%s.tmp", curl_handle->id);
 	curl = curl_handle->curl;
 
 	fptr = fopen(filename, "w+");
+	if (fptr == NULL) {
+		errcode = errno;
+		write_log(0, "IO Error in %s. Code %d, %s\n",
+			__func__, errcode, strerror(errcode));
+		return -1;
+	}
 	chunk = NULL;
 
 	sprintf(auth_url, "%s://%s/auth/v1.0", SWIFT_PROTOCOL, swift_url);
@@ -275,27 +335,36 @@ int hcfs_get_auth_swift(char *swift_user, char *swift_pass, char *swift_url,
 	curl_easy_setopt(curl, CURLOPT_CUSTOMREQUEST, NULL);
 	curl_easy_setopt(curl, CURLOPT_HTTPHEADER, chunk);
 
-	res = curl_easy_perform(curl);
+	HTTP_PERFORM_RETRY(curl);
 
 	if (res != CURLE_OK) {
 		fprintf(stderr, "failed %s\n", curl_easy_strerror(res));
 		fclose(fptr);
 		unlink(filename);
 		curl_slist_free_all(chunk);
-
-		return -1;
+			return -1;
 	}
 
-	ret_val = parse_auth_header(fptr);
+	ret_val = parse_swift_auth_header(fptr);
 
 	/*TODO: add retry routines somewhere for failed attempts*/
 
+	if (ret_val < 0) {
+		fclose(fptr);
+		unlink(filename);
+		curl_slist_free_all(chunk);
+		return ret_val;
+	}
+
 	fclose(fptr);
-	unlink(filename);
+	UNLINK(filename);
 
 	curl_slist_free_all(chunk);
 
-	return ret_val;
+	return 0;
+
+errcode_handle:
+	return -1;
 }
 
 /************************************************************************
@@ -431,10 +500,11 @@ size_t read_file_function(void *ptr, size_t size, size_t nmemb,
 {
 	/*TODO: Consider if it is possible for the actual file size to be
 		smaller than object size due to truncating*/
-	size_t total_size;
 	FILE *fptr;
 	size_t actual_to_read;
 	object_put_control *put_control;
+	int errcode;
+	size_t ret_size;
 
 	put_control = (object_put_control *) put_control1;
 
@@ -447,10 +517,12 @@ size_t read_file_function(void *ptr, size_t size, size_t nmemb,
 	else
 		actual_to_read = size * nmemb;
 
-	total_size = fread(ptr, 1, actual_to_read, fptr);
-	put_control->remaining_size -= total_size;
+	FREAD(ptr, 1, actual_to_read, fptr);
+	put_control->remaining_size -= ret_size;
 
-	return total_size;
+	return ret_size;
+errcode_handle:
+	return 0;
 }
 
 /************************************************************************
@@ -470,7 +542,7 @@ int hcfs_swift_list_container(CURL_HANDLE *curl_handle)
 	FILE *swift_list_header_fptr, *swift_list_body_fptr;
 	CURL *curl;
 	char header_filename[100], body_filename[100];
-	int ret_val;
+	int ret_val, num_retries, errcode;
 
 	sprintf(header_filename, "/run/shm/swiftlisthead%s.tmp",
 							curl_handle->id);
@@ -478,7 +550,21 @@ int hcfs_swift_list_container(CURL_HANDLE *curl_handle)
 	curl = curl_handle->curl;
 
 	swift_list_header_fptr = fopen(header_filename, "w+");
+	if (swift_list_header_fptr == NULL) {
+		errcode = errno;
+		write_log(0, "IO error in %s. Code %d, %s\n", __func__,
+			errcode, strerror(errcode));
+		return -1;
+	}
+
 	swift_list_body_fptr = fopen(body_filename, "w+");
+	if (swift_list_body_fptr == NULL) {
+		errcode = errno;
+		write_log(0, "IO error in %s. Code %d, %s\n", __func__,
+			errcode, strerror(errcode));
+		fclose(swift_list_header_fptr);
+		return -1;
+	}
 
 	chunk = NULL;
 
@@ -500,7 +586,8 @@ int hcfs_swift_list_container(CURL_HANDLE *curl_handle)
 	curl_easy_setopt(curl, CURLOPT_VERBOSE, 0L);
 	curl_easy_setopt(curl, CURLOPT_CUSTOMREQUEST, NULL);
 	curl_easy_setopt(curl, CURLOPT_HTTPHEADER, chunk);
-	res = curl_easy_perform(curl);
+
+	HTTP_PERFORM_RETRY(curl);
 
 	if (res != CURLE_OK) {
 		fprintf(stderr, "failed %s\n", curl_easy_strerror(res));
@@ -512,7 +599,7 @@ int hcfs_swift_list_container(CURL_HANDLE *curl_handle)
 		return -1;
 	}
 
-	ret_val = parse_list_header(swift_list_header_fptr);
+	ret_val = parse_swift_list_header(swift_list_header_fptr);
 
 	if ((ret_val >= 200) && (ret_val < 300))
 		dump_list_body(swift_list_body_fptr);
@@ -549,6 +636,7 @@ int hcfs_swift_put_object(FILE *fptr, char *objname, CURL_HANDLE *curl_handle)
 	CURL *curl;
 	char header_filename[100];
 	int ret_val;
+	int num_retries;
 
 	sprintf(header_filename, "/run/shm/swiftputhead%s.tmp",
 							curl_handle->id);
@@ -591,14 +679,14 @@ int hcfs_swift_put_object(FILE *fptr, char *objname, CURL_HANDLE *curl_handle)
 
 	curl_easy_setopt(curl, CURLOPT_URL, container_string);
 	curl_easy_setopt(curl, CURLOPT_HTTPHEADER, chunk);
-	res = curl_easy_perform(curl);
+
+	HTTP_PERFORM_RETRY(curl);
 
 	if (res != CURLE_OK) {
 		fprintf(stderr, "failed %s\n", curl_easy_strerror(res));
 		fclose(swift_list_header_fptr);
 		unlink(header_filename);
 		curl_slist_free_all(chunk);
-
 		return -1;
 	}
 
@@ -631,6 +719,7 @@ int hcfs_swift_get_object(FILE *fptr, char *objname, CURL_HANDLE *curl_handle)
 	CURL *curl;
 	char header_filename[100];
 	int ret_val;
+	int num_retries;
 
 	sprintf(header_filename, "/run/shm/swiftgethead%s.tmp",
 							curl_handle->id);
@@ -669,7 +758,7 @@ int hcfs_swift_get_object(FILE *fptr, char *objname, CURL_HANDLE *curl_handle)
 
 	curl_easy_setopt(curl, CURLOPT_URL, container_string);
 	curl_easy_setopt(curl, CURLOPT_HTTPHEADER, chunk);
-	res = curl_easy_perform(curl);
+	HTTP_PERFORM_RETRY(curl);
 
 	if (res != CURLE_OK) {
 		fprintf(stderr, "failed %s\n", curl_easy_strerror(res));
@@ -702,11 +791,11 @@ int hcfs_swift_delete_object(char *objname, CURL_HANDLE *curl_handle)
 	CURLcode res;
 	char container_string[200];
 	char delete_command[10];
-
 	FILE *swift_list_header_fptr;
 	CURL *curl;
 	char header_filename[100];
 	int ret_val;
+	int num_retries;
 
 	sprintf(header_filename, "/run/shm/swiftdeletehead%s.tmp",
 							curl_handle->id);
@@ -734,7 +823,7 @@ int hcfs_swift_delete_object(char *objname, CURL_HANDLE *curl_handle)
 	curl_easy_setopt(curl, CURLOPT_URL, container_string);
 	curl_easy_setopt(curl, CURLOPT_HTTPHEADER, chunk);
 
-	res = curl_easy_perform(curl);
+	HTTP_PERFORM_RETRY(curl);
 	if (res != CURLE_OK) {
 		fprintf(stderr, "failed %s\n", curl_easy_strerror(res));
 		fclose(swift_list_header_fptr);
@@ -865,13 +954,13 @@ int hcfs_S3_list_container(CURL_HANDLE *curl_handle)
 	FILE *S3_list_header_fptr, *S3_list_body_fptr;
 	CURL *curl;
 	char header_filename[100], body_filename[100];
-
 	unsigned char date_string[100];
 	char date_string_header[100];
 	unsigned char AWS_auth_string[200];
 	unsigned char S3_signature[200];
 	unsigned char resource[200];
 	int ret_val;
+	int num_retries;
 
 	sprintf(header_filename, "/run/shm/S3listhead%s.tmp", curl_handle->id);
 	sprintf(body_filename, "/run/shm/S3listbody%s.tmp", curl_handle->id);
@@ -909,7 +998,7 @@ int hcfs_S3_list_container(CURL_HANDLE *curl_handle)
 	curl_easy_setopt(curl, CURLOPT_CUSTOMREQUEST, NULL);
 	curl_easy_setopt(curl, CURLOPT_VERBOSE, 1L);
 	curl_easy_setopt(curl, CURLOPT_HTTPHEADER, chunk);
-	res = curl_easy_perform(curl);
+	HTTP_PERFORM_RETRY(curl);
 
 	if (res != CURLE_OK) {
 		fprintf(stderr, "failed %s\n", curl_easy_strerror(res));
@@ -1171,7 +1260,7 @@ int hcfs_S3_put_object(FILE *fptr, char *objname, CURL_HANDLE *curl_handle)
 	unsigned char S3_signature[200];
 	int ret_val;
 	unsigned char resource[200];
-
+	int num_retries;
 
 	sprintf(header_filename, "/run/shm/s3puthead%s.tmp", curl_handle->id);
 	sprintf(resource, "%s/%s", S3_BUCKET, objname);
@@ -1225,7 +1314,7 @@ int hcfs_S3_put_object(FILE *fptr, char *objname, CURL_HANDLE *curl_handle)
 
 	curl_easy_setopt(curl, CURLOPT_URL, container_string);
 	curl_easy_setopt(curl, CURLOPT_HTTPHEADER, chunk);
-	res = curl_easy_perform(curl);
+	HTTP_PERFORM_RETRY(curl);
 
 	if (res != CURLE_OK) {
 		fprintf(stderr, "failed %s\n", curl_easy_strerror(res));
@@ -1265,6 +1354,7 @@ int hcfs_S3_get_object(FILE *fptr, char *objname, CURL_HANDLE *curl_handle)
 	CURL *curl;
 	char header_filename[100];
 	int ret_val;
+	int num_retries;
 
 	unsigned char date_string[100];
 	char date_string_header[100];
@@ -1321,7 +1411,7 @@ int hcfs_S3_get_object(FILE *fptr, char *objname, CURL_HANDLE *curl_handle)
 
 	curl_easy_setopt(curl, CURLOPT_URL, container_string);
 	curl_easy_setopt(curl, CURLOPT_HTTPHEADER, chunk);
-	res = curl_easy_perform(curl);
+	HTTP_PERFORM_RETRY(curl);
 
 	if (res != CURLE_OK) {
 		fprintf(stderr, "failed %s\n", curl_easy_strerror(res));
@@ -1359,12 +1449,12 @@ int hcfs_S3_delete_object(char *objname, CURL_HANDLE *curl_handle)
 	CURL *curl;
 	char header_filename[100];
 	int ret_val;
-
 	unsigned char date_string[100];
 	char date_string_header[100];
 	unsigned char AWS_auth_string[200];
 	unsigned char S3_signature[200];
 	unsigned char resource[200];
+	int num_retries;
 
 	sprintf(header_filename, "/run/shm/s3deletehead%s.tmp",
 						curl_handle->id);
@@ -1404,7 +1494,7 @@ int hcfs_S3_delete_object(char *objname, CURL_HANDLE *curl_handle)
 
 	curl_easy_setopt(curl, CURLOPT_URL, container_string);
 	curl_easy_setopt(curl, CURLOPT_HTTPHEADER, chunk);
-	res = curl_easy_perform(curl);
+	HTTP_PERFORM_RETRY(curl);
 	if (res != CURLE_OK) {
 		fprintf(stderr, "failed %s\n", curl_easy_strerror(res));
 		fclose(S3_list_header_fptr);
